+++ src/main/kotlin/com/github/insanusmokrassar/AutoPostTelegramBot/plugins/publishers/PostPublisher.kt
@@ -79,7 +79,7 @@ class PostPublisher : Publisher {
                     logsChatId,
                     "Start post"
                 )
-            ).asMessage.let {
+            ).let {
                 messagesToDelete.add(it.chat.id to it.messageId)
             }